refact: removes config acessibility screen rules and liability
